@@ -17,80 +17,80 @@
   },
   "private": true,
   "dependencies": {
-    "@agm/core": "1.0.0-beta.5",
-    "@angular/animations": "7.0.2",
-    "@angular/cdk": "7.0.2",
-    "@angular/common": "7.0.2",
-    "@angular/compiler": "7.0.2",
-    "@angular/core": "7.0.2",
-    "@angular/forms": "7.0.2",
-    "@angular/http": "7.0.2",
-    "@angular/material": "7.0.2",
-    "@angular/platform-browser": "7.0.2",
-    "@angular/platform-browser-dynamic": "7.0.2",
-    "@angular/platform-server": "7.0.2",
-    "@angular/router": "7.0.2",
-    "ajv": "6.3.0",
-    "angular-calendar": "0.26.4",
-    "animate.css": "3.5.2",
-    "arrive": "2.3.1",
-    "bootstrap": "4.1.3",
-    "bootstrap-material-design": "4.1.1",
+    "@agm/core": "1.0.0-beta.6",
+    "@angular/animations": "8.0.0",
+    "@angular/cdk": "8.0.0",
+    "@angular/common": "8.0.0",
+    "@angular/compiler": "8.0.0",
+    "@angular/core": "8.0.0",
+    "@angular/forms": "8.0.0",
+    "@angular/http": "7.2.15",
+    "@angular/material": "8.0.0",
+    "@angular/platform-browser": "8.0.0",
+    "@angular/platform-browser-dynamic": "8.0.0",
+    "@angular/platform-server": "8.0.0",
+    "@angular/router": "8.0.0",
+    "ajv": "6.10.0",
+    "angular-calendar": "0.27.9",
+    "animate.css": "3.7.2",
+    "arrive": "2.4.1",
+    "bootstrap": "^4.3.1",
+    "bootstrap-material-design": "4.1.2",
     "bootstrap-notify": "3.1.3",
     "bootstrap-rtl": "3.3.4",
-    "bootstrap-select": "1.12.4",
+    "bootstrap-select": "1.13.10",
     "bootstrap-tagsinput": "0.7.1",
     "chartist": "0.9.8",
     "chartist-plugin-zoom": "0.4.1",
-    "core-js": "2.4.1",
+    "core-js": "3.1.3",
     "datatables": "1.10.18",
     "datatables.net-bs4": "1.10.19",
-    "datatables.net-responsive": "2.1.1",
+    "datatables.net-responsive": "2.2.3",
     "fullcalendar": "3.4.0",
-    "googleapis": "19.0.0",
+    "googleapis": "40.0.0",
     "hammerjs": "2.0.8",
     "jasny-bootstrap": "3.1.3",
-    "jquery": "3.2.1",
+    "jquery": "^3.4.1",
     "jquery-validation": "1.17.0",
     "jqvmap": "1.5.1",
-    "moment": "2.22.2",
-    "moment-timezone": "0.4.1",
-    "ng2-nouislider": "1.7.13",
+    "moment": "2.24.0",
+    "moment-timezone": "0.5.25",
+    "ng2-nouislider": "1.8.2",
     "ng2-select": "2.0.0",
     "ngx-chips": "1.9.8",
     "nouislider": "9.2.0",
     "perfect-scrollbar": "1.1.0",
-    "popper.js": "1.14.4",
-    "rxjs": "6.1.0",
-    "rxjs-compat": "6.1.0",
+    "popper.js": "1.15.0",
+    "rxjs": "6.5.2",
+    "rxjs-compat": "6.5.2",
     "sweetalert2": "7.22.2",
     "twitter-bootstrap-wizard": "1.2.0",
-    "uglify-js": "3.3.28",
+    "uglify-js": "3.6.0",
     "validate": "3.0.1",
     "web-animations-js": "2.3.1",
-    "zone.js": "0.8.26"
+    "zone.js": "0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.6.3",
-    "@angular/cli": "7.0.4",
-    "@angular/compiler-cli": "7.0.2",
-    "@types/bootstrap": "3.3.32",
-    "@types/chartist": "0.9.34",
-    "@types/jasmine": "2.5.38",
-    "@types/jquery": "1.10.31",
-    "@types/node": "6.0.73",
-    "codelyzer": "4.3.0",
-    "jasmine-core": "2.5.2",
-    "jasmine-spec-reporter": "3.2.0",
-    "karma": "3.1.1",
-    "karma-chrome-launcher": "2.0.0",
-    "karma-cli": "1.0.1",
-    "karma-coverage-istanbul-reporter": "0.2.0",
-    "karma-jasmine": "1.1.0",
-    "karma-jasmine-html-reporter": "0.2.2",
-    "protractor": "5.4.1",
-    "ts-node": "2.0.0",
-    "tslint": "5.9.1",
-    "typescript": "3.1.6"
+    "@angular-devkit/build-angular": "0.800.1",
+    "@angular/cli": "8.0.1",
+    "@angular/compiler-cli": "8.0.0",
+    "@types/bootstrap": "4.3.0",
+    "@types/chartist": "0.9.46",
+    "@types/jasmine": "3.3.13",
+    "@types/jquery": "3.3.29",
+    "@types/node": "12.0.5",
+    "codelyzer": "5.1.0",
+    "jasmine-core": "3.4.0",
+    "jasmine-spec-reporter": "4.2.1",
+    "karma": "4.1.0",
+    "karma-chrome-launcher": "2.2.0",
+    "karma-cli": "2.0.0",
+    "karma-coverage-istanbul-reporter": "2.0.5",
+    "karma-jasmine": "2.0.1",
+    "karma-jasmine-html-reporter": "1.4.2",
+    "protractor": "5.4.2",
+    "ts-node": "8.2.0",
+    "tslint": "5.17.0",
+    "typescript": "3.4.5"
   }
 }
